LayoutTests/editing/execCommand/insert-list-empty-div-expected.txt


<DIV id="div" contentEditable="true">
<UL>
<LI><selection-caret>
<BR>
</LI>
</UL>
</DIV>
